style(client-presence): rename `experimentalAccess.ts` to `getPresence.ts`
