+++ packages/framework/presence/src/getPresence.ts
