+++ packages/framework/presence/tsconfig.main.json
@@ -1,7 +1,7 @@
 {
 	"extends": "../../../common/build/build-common/tsconfig.node16.json",
 	"include": ["src/**/*"],
-	"exclude": ["src/experimentalAccess.ts", "src/index.ts", "src/test/**/*"],
+	"exclude": ["src/getPresence.ts", "src/index.ts", "src/test/**/*"],
 	"compilerOptions": {
 		"rootDir": "./src",
 		"outDir": "./lib",
